Enable Linux loader stub for universal payload
This patch enables a linux stub to be packed together with kernel
image. The stub will do HOB parsing and image loading for kernel and
then jump into kernel entrypoint.

To build this stub:
BuildLoader.py build_dsc -a ia32
               -p UniversalPayloadPkg\UniversalPayloadPkg.dsc -r
This stub needs to be provided to pack_payload script when the payload
type is specified as Linux type.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>